+++ tests/os/storage_layout_test.h
@@ -57,7 +57,7 @@ void StorageLayoutTest::testAddDirectory()
     Filename f1("/1");
     layout.addDirectory(f1);
     vector<FSBlock> blocks = layout.getFileMetaDataBlocks(f1);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)0, blocks[0]);
 }
 
@@ -67,12 +67,12 @@ void StorageLayoutTest::testAddFile()
     Filename f1("/1");
     layout.addFile(f1, 1024);
     vector<FSBlock> blocks = layout.getFileMetaDataBlocks(f1);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)0, blocks[0]);
 
     // Data blocks start at block 1000
     blocks = layout.getFileDataBlocks(f1, 0, 1024);
-    CPPUNIT_ASSERT_EQUAL(4u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)4, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1000, blocks[0]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[1]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1002, blocks[2]);
@@ -87,7 +87,7 @@ void StorageLayoutTest::testGetFileMetaDataBlocks()
     Filename f1("/1");
     layout.addDirectory(f1);    
     vector<FSBlock> blocks1 = layout.getFileMetaDataBlocks(f1);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks1.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks1.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)0, blocks1[0]);
 
     
@@ -95,7 +95,7 @@ void StorageLayoutTest::testGetFileMetaDataBlocks()
     Filename f2("/2");
     layout.addFile(f2, 0);
     vector<FSBlock> blocks2 = layout.getFileMetaDataBlocks(f2);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks2.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks2.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1, blocks2[0]);
 }
 
@@ -110,26 +110,26 @@ void StorageLayoutTest::testGetFileDataBlocks()
 
     // Test retrieving no blocks
     blocks = layout.getFileDataBlocks(f1, 0, 0);
-    CPPUNIT_ASSERT_EQUAL(0u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)0, blocks.size());
 
     // Test retrieving entire first block
     blocks = layout.getFileDataBlocks(f1, 0, 256);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1000, blocks[0]);
 
     // Test retrieving entire last block
     blocks = layout.getFileDataBlocks(f1, 1792, 256);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1007, blocks[0]);
 
     // Test retrieving entire 2nd block
     blocks = layout.getFileDataBlocks(f1, 256, 256);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[0]);
 
     // Test retrieving entire file
     blocks = layout.getFileDataBlocks(f1, 0, 2048);
-    CPPUNIT_ASSERT_EQUAL(8u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)8, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1000, blocks[0]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[1]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1002, blocks[2]);
@@ -141,28 +141,28 @@ void StorageLayoutTest::testGetFileDataBlocks()
 
     // Test a single block request in the center of a block
     blocks = layout.getFileDataBlocks(f1, 312, 4);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[0]);
 
     // Test a single block request at the beginning of a block
     blocks = layout.getFileDataBlocks(f1, 256, 12);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[0]);
 
     // Test a single block request at the end of a block
     blocks = layout.getFileDataBlocks(f1, 500, 11);
-    CPPUNIT_ASSERT_EQUAL(1u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[0]);
 
     // Test a small request that spans two blocks
     blocks = layout.getFileDataBlocks(f1, 248, 20);
-    CPPUNIT_ASSERT_EQUAL(2u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)2, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1000, blocks[0]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[1]);
 
     // Test a request with partial blocks at the beginning and end
     blocks = layout.getFileDataBlocks(f1, 248, 312);
-    CPPUNIT_ASSERT_EQUAL(3u, blocks.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)3, blocks.size());
     CPPUNIT_ASSERT_EQUAL((FSBlock)1000, blocks[0]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1001, blocks[1]);
     CPPUNIT_ASSERT_EQUAL((FSBlock)1002, blocks[2]);
